관리자 페이지 검색 기능 추가중

+++ src/main/java/shop/mtcoding/blogv3/service/AdminService.java
@@ -10,7 +10,9 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.mtcoding.blogv3.dto.user.UserReqDto.LoginReqDto;
 import shop.mtcoding.blogv3.handler.ex.CustomApiException;
 import shop.mtcoding.blogv3.handler.ex.CustomException;
+import shop.mtcoding.blogv3.model.Board;
 import shop.mtcoding.blogv3.model.BoardRepository;
+import shop.mtcoding.blogv3.model.Reply;
 import shop.mtcoding.blogv3.model.ReplyRepository;
 import shop.mtcoding.blogv3.model.User;
 import shop.mtcoding.blogv3.model.UserRepository;
@@ -52,6 +54,10 @@ public class AdminService {
             throw new CustomApiException("삭제 권한이 없습니다.", HttpStatus.FORBIDDEN);
         }
 
+        Reply reply = replyRepository.findById(id);
+        if (reply == null) {
+            throw new CustomApiException("댓글이 존재하지 않습니다.");
+        }
         int result = replyRepository.deleteById(id);
 
         if (result != 1) {
@@ -67,8 +73,34 @@ public class AdminService {
             throw new CustomApiException("삭제 권한이 없습니다.", HttpStatus.FORBIDDEN);
         }
 
+        Board board = boardRepository.findById(id);
+
+        if (board == null) {
+            throw new CustomApiException("게시글이 존재하지 않습니다.");
+        }
+
         int result = boardRepository.deleteById(id);
 
+        if (result != 1) {
+            throw new CustomApiException("삭제 실패", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @Transactional
+    public void UserDelete(int userId) {
+        User principal = (User) session.getAttribute("principal");
+
+        if (!principal.getRoll().equals("ADMIN")) {
+            throw new CustomApiException("삭제 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+
+        User user = userRepository.findById(userId);
+
+        if (user == null) {
+            throw new CustomApiException("계정이 존재하지 않습니다.");
+        }
+
+        int result = userRepository.deleteById(userId);
         if (result != 1) {
             throw new CustomApiException("삭제 실패", HttpStatus.INTERNAL_SERVER_ERROR);
         }
